fix: bug in indexing of week
