+++ src/lib/session-availible-table.ts
@@ -1,3 +1,5 @@
+export const dayNames = ["SUN", "MON", "TUE", "WED", "THU", "FRI", "SAT"];
+
 export const generateWeekDays = (startMonday?: Date) => {
   let startDate: Date;
 
@@ -43,4 +45,15 @@ export const generateTimeSlots = () => {
     }
   }
   return slots;
+};
+
+export const getCurrentWeekMonday = (currentWeek: number = 0) => {
+  const today = new Date();
+  const currentMonday = new Date(today);
+  const dayOfWeek = today.getDay();
+  const daysToMonday = dayOfWeek === 0 ? -6 : 1 - dayOfWeek;
+
+  currentMonday.setDate(today.getDate() + daysToMonday);
+  currentMonday.setDate(currentMonday.getDate() + currentWeek * 7);
+  return currentMonday;
 };
